Extracted compatibility module to work without AioModel subclassing + tests
