+++ tests/test_compat.py
@@ -1,0 +1,41 @@
+import uuid
+
+from tests.conftest import all_dbs
+from tests.models import CompatTestModel
+
+
+@all_dbs
+async def test_create_select_compat_mode(manager):
+    obj1 = await manager.create(CompatTestModel, text="Test 1")
+    obj2 = await manager.create(CompatTestModel, text="Test 2")
+    query = CompatTestModel.select().order_by(CompatTestModel.text)
+    result = await manager.execute(query)
+    assert list(result) == [obj1, obj2]
+
+
+@all_dbs
+async def test_update_compat_mode(manager):
+    obj_draft = await manager.create(CompatTestModel, text="Draft 1")
+    obj_draft.text = "Final result"
+    await manager.update(obj_draft)
+    obj = await manager.get(CompatTestModel, id=obj_draft.id)
+    assert obj.text == "Final result"
+
+
+@all_dbs
+async def test_count_compat_mode(manager):
+    obj = await manager.create(CompatTestModel, text="Unique title %s" % uuid.uuid4())
+    search = CompatTestModel.select().where(CompatTestModel.text == obj.text)
+    count = await manager.count(search)
+    assert count == 1
+
+
+@all_dbs
+async def test_delete_compat_mode(manager):
+    obj = await manager.create(CompatTestModel, text="Expired item %s" % uuid.uuid4())
+    search = CompatTestModel.select().where(CompatTestModel.id == obj.id)
+    count_before = await manager.count(search)
+    assert count_before == 1
+    await manager.delete(obj)
+    count_after = await manager.count(search)
+    assert count_after == 0
